feat : refactor admin panel and add view HCP and approve card components

+++ components/base/admin-dashboard/data-table/healthcare-centers-table/actions.tsx
@@ -11,6 +11,17 @@ import { User } from "@prisma/client";
 import { CheckIcon, EyeOpenIcon, TrashIcon } from "@radix-ui/react-icons";
 import { toast } from "sonner";
 
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import { Spinner } from "@/components/ui/spinner";
 
@@ -30,11 +41,19 @@ export function HealthcareCentersDataTableActions({
   healthcareCenter,
 }: HealthcareCentersDataTableActionsProps) {
   const router = useRouter();
+  const healthcareCenterEmail = healthcareCenter!.user.email;
+  const healthcareCenterName = healthcareCenter!.user.name;
   const [isPending, startTransition] = React.useTransition();
 
   const handleVerify = (id: string) => {
     startTransition(() => {
-      verifyHealthcareCenter(id).then((data) => {
+      verifyHealthcareCenter(
+        id,
+        healthcareCenterEmail,
+        "oladoc-customer-service",
+        "From Oladoc customer service team!",
+        healthcareCenterName,
+      ).then((data) => {
         if (data?.error) {
           toast.error(data.error);
           return;
@@ -48,7 +67,13 @@ export function HealthcareCentersDataTableActions({
 
   const handleDelete = (id: string) => {
     startTransition(() => {
-      deleteHealthcareCenter(id).then((data) => {
+      deleteHealthcareCenter(
+        id,
+        healthcareCenterEmail,
+        "oladoc-customer-service",
+        "From Oladoc customer service team!",
+        healthcareCenterName,
+      ).then((data) => {
         if (data?.error) {
           toast.error(data.error);
           return;
@@ -67,7 +92,7 @@ export function HealthcareCentersDataTableActions({
         variant="outline"
         className="gap-1"
         onClick={() =>
-          router.push(`/admin/healthcare-providers/${healthcareCenter.id}`)
+          router.push(`/admin/healthcare-centers/${healthcareCenter.id}`)
         }
       >
         <EyeOpenIcon className="h-4 w-4" />
@@ -83,16 +108,38 @@ export function HealthcareCentersDataTableActions({
         {isPending ? <Spinner /> : <CheckIcon className="h-4 w-4" />}
         <span>Verify</span>
       </Button>
-      <Button
-        size="sm"
-        variant="destructive"
-        className="gap-1"
-        onClick={() => handleDelete(healthcareCenter.id)}
-        disabled={isPending}
-      >
-        {isPending ? <Spinner /> : <TrashIcon className="h-4 w-4" />}
-        <span>Delete</span>
-      </Button>
+
+      <AlertDialog>
+        <AlertDialogTrigger asChild>
+          <Button
+            size="sm"
+            variant="destructive"
+            className="gap-1"
+            disabled={isPending}
+          >
+            {isPending ? <Spinner /> : <TrashIcon className="h-4 w-4" />}Delete
+          </Button>
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>
+              Are you sure you want to delete this account?
+            </AlertDialogTitle>
+            <AlertDialogDescription>
+              This action cannot be undone. This will permanently delete this
+              account and remove the data from our servers.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel className="mt-0">Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={() => handleDelete(healthcareCenter.id)}
+            >
+              Continue
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     </div>
   );
 }
